+++ Source/Serialization/TypeSerializers/TransactionSerializer.swift
@@ -6,7 +6,7 @@
 
 import Foundation
 
- /// https://github.com/nervosnetwork/ckb/blob/develop/util/types/schemas/ckb.mol#L69
+/// https://github.com/nervosnetwork/ckb/blob/develop/util/types/schemas/blockchain.mol#L55-L67
 public final class TransactionSerializer: TableSerializer<Transaction> {
     public required init(value: Transaction) {
         let hexStringsToArrayOfBytes: ([HexString]) -> [[Byte]] = { strings in
@@ -20,7 +20,7 @@ public final class TransactionSerializer: TableSerializer<Transaction> {
                 FixVecSerializer<[Byte], Byte32Serializer>(value: hexStringsToArrayOfBytes(value.headerDeps)),
                 FixVecSerializer<CellInput, CellInputSerializer>(value: value.inputs),
                 DynVecSerializer<CellOutput, CellOutputSerializer>(value: value.outputs),
-                DynVecSerializer<[Byte], FixVecSerializer<Byte, ByteSerializer>>(value: hexStringsToArrayOfBytes(value.outputsData))
+                DynVecSerializer<[Byte], BytesSerializer>(value: hexStringsToArrayOfBytes(value.outputsData))
             ]
         )
     }
@@ -35,5 +35,51 @@ public extension Transaction {
     func computeHash() -> H256 {
         let serialized = serialize()
         return Blake2b().hash(bytes: serialized)
+    }
+}
+
+// - MARK: Size & Fee
+
+final class TransactionPlusWitnessesSerializer: TableSerializer<Transaction> {
+    required init(value: Transaction) {
+        let witnesses: [[Byte]]
+        if value.witnesses.isEmpty {
+            witnesses = value.serializeWitnessArgs()
+        } else {
+            witnesses = value.witnesses.map { Data(hex: $0).bytes }
+        }
+
+        super.init(
+            value: value,
+            fieldSerializers: [
+                TransactionSerializer(value: value),
+                DynVecSerializer<[Byte], BytesSerializer>(value: witnesses)
+            ]
+        )
+    }
+}
+
+extension Transaction {
+    var serializedSizeInBlock: Int {
+        let serializer = TransactionPlusWitnessesSerializer(value: self)
+        let serializedTxOffsetBytesize = 4 // 4 bytes for the tx offset cost with molecule array (transactions)
+        return serializer.serialize().count + serializedTxOffsetBytesize
+    }
+
+    public func fee(rate: UInt64) -> Capacity {
+        return Transaction.fee(for: serializedSizeInBlock, with: rate)
+    }
+
+    func serializeWitnessArgs() -> [[Byte]] {
+        return unsignedWitnesses.map { $0.serialize() }
+    }
+
+    /// Calulate fee based on transaction size and fee rate.
+    /// - Parameter txSize: Bytesize of the serialized tx in a block
+    /// - Parameter rate: Fee rate, shannons for 1KB
+    public static func fee(for txSize: Int, with rate: UInt64) -> Capacity {
+        let base = UInt64(txSize) * rate
+        let result = base / 1000
+        return base % 1000 == 0 ? result : result + 1
     }
 }