+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [v0.24.0](https://github.com/nervosnetwork/ckb-sdk-swift/compare/v0.23.1...v0.24.0) (2019-11-02)
+
+
+### Features
+
+* Calculate transaction fee ([73dae6e](https://github.com/nervosnetwork/ckb-sdk-swift/commit/73dae6e))
+* Implement estimate_fee_rate RPC ([9e43807](https://github.com/nervosnetwork/ckb-sdk-swift/commit/9e43807))
+* Implement tx size calculation ([b4bac55](https://github.com/nervosnetwork/ckb-sdk-swift/commit/b4bac55))
+* Implement WitnessArgs and its serializer ([f052a04](https://github.com/nervosnetwork/ckb-sdk-swift/commit/f052a04))
+* Let WitnessArgs support both parsed witness and raw data ([15dfe74](https://github.com/nervosnetwork/ckb-sdk-swift/commit/15dfe74))
+* Update tx signing ([499978e](https://github.com/nervosnetwork/ckb-sdk-swift/commit/499978e))
+
+
+
 # [v0.23.1](https://github.com/nervosnetwork/ckb-sdk-swift/compare/v0.23.0...v0.23.1) (2019-10-19)